Texel: Add contempt UCI parameter.
Add UCI parameters for game playing contempt and for analysis contempt.
Analysis contempt is relative to the white player and used when
analyzing a position. Game playing contempt is relative to the side to
move and used when not in analyze mode.

Does not help much against weaker opponents (~5 elo best observed case),
but does help quite a big against stronger opponents. (+40 elo observed
against stockfish 8 with contempt = -75).