Make ABDADA work for deterministic search
The ABDADA-like algorithm performs read-modify-write operations on the
transposition table when checking and updating the "busy" bit. Make
these read-modify-write operations atomic with respect to other
threads to prevent all helper threads from first all reading the same
busy bit to conclude that it is not set, then all setting that same
busy bit to true and start searching the corresponding move.