Fix non-determinism when searching for more than 1 second
The Search::notifyStats() function is called based on elapsed time
(about once per second) instead of node counts. This means that it
must not interact with the ThreadOrder class, as that would introduce
non-determinism.

This change means that the reported "hashFull" value will be subject
to timing, but that was already the case before since the times when
Search::notifyStats() was called was non-deterministic.