In proof game search, print some info before starting the search
Before starting the search, print information about the required piece
count changes (because of trades and promotions) to get from the
initial position to the goal position.

This makes it easier to manually construct a partial game suitable for
the "-ipgn" option when the problem is too hard to be solved
automatically.