Create a reverse move generator
For a given position, the reverse move generator calculates a list of
moves and corresponding undo information, such that for all legal
positions that lead to the target position by playing a suitable
legal move, that legal move is included in the list.

It is too hard to make sure the list never contains moves
corresponding to illegal previous positions, because that would
require determining if an arbitrary position is legal. Some easy
cases are handled though:

* Moves that are invalid because the resulting piece counts (taking
  promotions and captures into account) are impossible.
* Moves that are invalid because king capture was possible before the
  move.
* Moves that are invalid because the en passant square does not match
  a double pawn push.