Better handling of bishops that cannot move after promotion
If a bishop is trapped on a pawn promotion square in the goal position
and the bishop was not there in the initial position, it follows that
a pawn must have promoted to a bishop on the corresponding file. This
requires two things:

1. The pawn configuration on the file must allow for at least one
   promotion. This was already handled before this commit.
2. The total number of bishop promotions is at least one. (Or more
   generally, at least equal to the number of files where this
   situation occurs.) This is implemented in this commit.

This improvement causes the position in

  https://github.com/tromp/ChessPositionRanking/issues/221

to be identified as illegal.