Fix rook/queen promotion next to unmovable opponent king
The following position was incorrectly classified as illegal by the
proof kernel search:

    r1bRkbnr/pppp1ppp/8/8/8/8/PPPP1PPP/RNBQKBNR b KQkq - 0 1

It is however legal as shown by the proof game:

    e4 Nc6 e5 Nf6 exf6 Nb4 fxe7 Nd5 Qe2 Nf6 Qd1 Ng8 exd8R+

The problem was that the search contains code to forbid promotion to
rook/queen on a square next to a king that is not allowed to move
because it needs to retain its castling rights. The just promoted
piece would then have to be captured on the next move since the king
cannot move away to evade the check, thereby making the promotion
pointless. However, this is not true if the promotion move is the very
last move in the proof game.

The code is now changed so that rook/queen promotion is allowed if
there is an opponent rook/queen next to the king in the goal position.