Only cache failed results in proof kernel search
Only store states known to not lead to a goal position in the cache.
This does not increase search time because the search is stopped as
soon as the first solution is found.

This change will be important when adding CSP verification to found
proof kernels, because state S1 may lead to goal G1 though path P1,
but P1 fails CSP verification. However it is possible that another
path P2 leads to state S1 and G1, but P2 passes CSP verification.
Caching S1 would lead to path P2 not being found.