Don't stop proof kernel search until all captures are made
In proof kernel search, don't treat a state as a goal state unless it
has the same number of pieces as the goal position. This will simplify
ensuring that a position is illegal if no CSP verification succeeds.

This change requires also generating "piece takes piece" moves during
search, otherwise a position where no pawn can move and a non-pawn
piece needs to be captured will be incorrectly classified as illegal.
It is enough to consider such moves at the end of the move sequence
though, because they don't interfere with moves involving pawns.

This change also speeds up classification of random positions because
there are fewer calls to the expensive ProofKernel::isGoal() function.