Implement extended proof kernel search
An extended proof kernel is like a normal proof kernel but with the
additional constraints that captures need to happen on specific
squares.

Some proof kernels don't have a corresponding extended proof kernel
because there is no way to make all captures happen on correct squares
while respecting chess rules like "pawns cannot move backwards".

When the regular proof kernel search finds a valid proof kernel, it
calls the extended proof kernel search to see if it has a
corresponding solution. If not, the regular search continues. The
problem of finding an extended proof kernel is a relaxation of the
problem of finding a sequence of normal chess moves to get to a state
where the required pawn promotions can be made. Therefore, if there is
no extended proof kernel for a goal position, it is not reachable from
the start position.