Better handling of pawn promotion moves
There are rows > 7 and < 0 that are used for allowing already promoted
pawns to move to leave room for other pawns. If a pawn moves from a
non-promoted row (1-7) to an "out of board" row, display that move as
a regular promotion move.

This probably currently never happens because the CSP solver always
finds the smallest/largest valid value when looking for a solution.