Better pruning in proof kernel search
A move in the proof kernel space can at best add one pawn to one
column. Therefore, if the total number of pawns needed to be added for
one color is larger than the number of remaining captures of the other
color, it is not possible to reach a goal state.

The above can be further improved by taking into account how many files
away the closest available pawn is.