Increase max number of supported pawns in one column
In the proof kernel search, the code now supports up to 7 pawns in one
column (file). A compile time constant decides the maximum, so it can
easily be increased further if needed. It is not desirable to set the
value too large, because there are tables of size 2 ^ (maxPawns + 1)
that need to be initialized for each position to analyze.

This change makes it possible to analyze the following position:

    1BQqN1K1/4pp2/rqq1PQ2/1p5p/1P5P/p3b3/1rkB2P1/1Q1b2rQ b - - 1 3