In proof game search, improve calculation of blocked squares
A pawn is blocked (i.e. not possible to move) if the pawn is present
at the same square in the goal position and there is no other pawn
that can reach the square.

This calculation now also considers the total number of available
captures when deciding if another pawn can reach the square.