Revert "In proof game search, improve heuristic function"
This reverts most of commit 66984fa339eda23abbe830a2cd4350e706eb9fa4,
since it is wrong for the case where no captures are available but
pawn promotions are. Since the heuristic did not help much anyway it
is not worth fixing.

Some tests are kept and a new test demonstrating why the code was
wrong has been added.