Optional non-admissible heuristic in proof game search
Add command line option to "texelutil proofgame" to use a
non-admissible heuristic for the A* search. This heuristic takes
pieces blocking other pieces into account, and is sometimes able to
solve problems quickly that the default heuristic find very hard to
solve.

There is no way to use this option for "texelutil proofgame -f" yet.