Use non-admissible heuristic in "texelutil proofgame -f"
If the standard proof game search fails in "proofgame -f" try to find
a path from the best attempt to the goal using the non-admissible
heuristic. In some cases the standard search leads to a position where
only one or a few pieces are not on their target squares, but there is
a large number of pieces blocking their paths. The non-admissible
heuristic can then often find a solution quickly.