Use multiple threads in "texelutil proofgame -f"
The result is still deterministic. Logging to std::cerr is delayed so
that the same information is logged in the same order as when using a
single thread.

The number of search threads defaults to the number of hardware
threads in the computer unless explicitly set using "-j nThreads".
For example:

  cat fens | texelutil -j 4 proofgame -f -o out 2>debug | tee status

Result on an 8 core computer for the 10k sample set:

legal: 527 path: 0 kernel: 0 fail: 11 illegal: 9462 time: 2034.083