Improve rejection of possible last moves in proof game search
In proof game search, when a possible last move is a capture, the code
used to invoke proof kernel code to try to prove the previous position
illegal. This behavior is kept, but additionally it also invokes
itself recursively with a small "max allowed search nodes" value to
try to prove the previous position illegal by piece reachability
analysis.

Also in proof game "retry" mode, redo also the proof kernel
calculation so that the improvement in this commit will be used.