In proof game search, handle positions where last move is a capture
If the possible last moves for a goal position are all captures and
there is more than one capture available, assume the first found valid
capture is actually the last move in the corresponding proof game.
This makes the proof kernel and the corresponding "path" not contain
the final capture, which is necessary for the proof game search to
succeed.