Better detection of deadlocks in proof game search
In proof game search, if there is a set of pieces where no piece can
move unless a different piece in the set is moved first, it means that
none of the pieces in the set can move. Use this fact to detect when
the goal position cannot be reached from a given position.

This code replaces the old computeKingPawnsTrapBlocked() code that did
the same type of analysis but only for one special case.