In proof game search, better handling of forced last moves
Detect the situation where a position is illegal because there is no
valid second-to-last move, even though there are multiple quiet valid
last moves. This analysis is only performed if the side to move before
the last move is in check, because that is where the problem usually
occurs.

This causes the following positions from the 1M sample to be
classified as illegal:

5r2/P1r1R1nR/kb2Bb2/1n1qpnQq/b1BB1B2/KN2P1np/n2N4/4N3 w - - 0 1
NN1b1K2/pQ3Qn1/2qQ1q1N/7k/2B4n/Nnb2rQR/r3RPr1/1qB4b b - - 0 1
2QN1nqq/k3Pb1q/prp2b2/1N5R/B1P5/QK3Nrb/2N2R2/bB2r1B1 b - - 0 1
2rNK3/R1B4R/1bp1k1Bb/1nRb2R1/rNPP4/Q2QN1nq/n2n2q1/N7 b - - 0 1
3Q4/1Rq1r1Q1/2P2pbB/3n2PB/1NnB3Q/1k4r1/1PN3qR/qqNKb2q b - - 0 1