Move non-pawn pieces away to make other non-pawn move possible
Not finished. Causes segmentation fault (after search explosion) in
ProofGameTest.testFilter1b.

Causes problem in move sequence "bQd8-a8 wPb7xa8Q". Pieces between d8
and a8 are now moved away, which means the black king is in check
after the pawn promotion, but there is no code yet to avoid the
possible king capture.

The wrong piece can be captured if there are two capture promotions on
the same square, where the first captures a piece just moved to the
square, and the second is supposed to capture the original piece on
the square. The original piece is now moved away before the first
capture, and not moved back afterwards.