Add option to randomize proof kernel search
Add "texelutil proofgame -f" option "-rndkernel" to randomize the
search for a proof kernel. The idea is that if it seems too hard to
find a proof game corresponding to a proof kernel, a different proof
kernel for the same position might be easier to find a proof game for.