Use std::clog instead of std::cerr in proofgame filtering
Since C++11, cerr is tied to cout, which means that if
std::ios::sync_with_stdio(false) is used, two threads cannot safely
write to cout and cerr at the same time.

clog is like cerr except it is not tied to cout, so one thread can
safely write to clog while another thread writes to cout.

This fixes corrupted status reporting on cout when running

    cat indata | texelutil -j 1 proofgame -f -o out 2>debug