Add syncengine program
syncengine acts as a filter for a UCI engine, so that commands are not
sent to the engine while it is busy. Useful for testing a UCI engine
in a script. For example:

for i in $(seq 5) ; do echo -e "go depth 14\nquit" | ./syncengine ./texel | tail -n 2 | grep nps ; done