Handle sparse input to second NN layer, 110a7
In the matrix multiplication for the second layer, optimize for
sparseness of the first layer outputs.

Also permute current net using "torchutil quant -p" to make the first
layer outputs contain more consecutive zeros on average.

Not enabled for SSSE3 because it is slower than the non-sparse code,
at least in the initial position.