torchutil: Set unused layer 1 weights to 0 during quantization
Some rows in the first layer weights matrix are unused because they
correspond to illegal positions, like pawn on first/last rank or piece
on the same square as the king. Set those rows to zero in the weight
matrix.

This does not affect evaluation, but can lead to a better bounds for
the "L1 out" statistics, and also makes the network compress slightly
better.