torchutil: Compute better bounds for min/max layer 1 outputs
Consider each color separately and the fact that there are at most 15
non-king pieces for each color on the board.

This is not guaranteed to give better bounds than the old code, for
example if the values are very large for both colors on 15 squares,
but very small on the other squares.

For the current net the bounds become around 40% smaller. The largest
values come from a feature emulating a piece/square table, where the
largest values come from a queen and the smallest values come from a
queen of the other color.