Avoid expensive DTZ probes during search
Syzygy DTZ tables are one-sided, meaning that only data for either
"white to move" or "black to move" is stored in a table. If probing
for the non-existing side to move, the TB probing code does an
internal one-ply search, which is a lot more expensive than a single
probe. In this situation, it is better to give up so that the regular
search will instead go to the next ply and try the TB probe there,
where the side to move will agree with the TB file. This has two
benefits:

* Fewer TB probes are needed, because the alpha/beta interval in the
  main search is often smaller than the interval in the Syzygy probing
  code.

* The time it takes to stop the search is reduced, because checking
  the stop condition only happens in the main search, not when the
  Syzygy code does an internal one-ply search.