Stop sooner if running out of time when helper threads have results
If TB probes are very slow, it is possible that the main thread wants
to stop searching after the first TB probe. However, if there are many
helper threads searching, one of them may have a result before the
stop check is performed, causing a HelperThreadResult to be thrown
instead of StopSearch. If this happens repeatedly the engine could run
out of time.

The fix is to reorder the code so that the stop check happens before
the check for helper thread results.